==> code/1_inst_cleaning/00_diversity_institutions.ipynb <==
%config Completer.use_jedi = False

import pandas as pd
import pickle
from tqdm.notebook import tqdm
import numpy as np


from libs.LastNamesInference import LastNamesInference

address_clean = pd.read_pickle('/data/datasets/WOS/US/address_clean.p')

carnegie_df = pd.read_excel('../../data/carnegie/CCIHE2021-PublicData.xlsx',sheet_name='Data',
                           converters = {'unitid':str,'selindex':str})

carnegie_df = carnegie_df[['unitid','name','city','hbcu','hsi','womens','tribal','basic2021']].copy()

carnegie_df['retrieved']= carnegie_df.unitid.isin(set(address_clean.carnegie_id))
carnegie_df['R1'] = carnegie_df.basic2021 == 16

carnegie_df['basic2021'] = carnegie_df.basic2021.replace({0:"(Not classified)",
                                                1:"Associate's Colleges: High Transfer-High Traditional",
                                                2:"Associate's Colleges: High Transfer-Mixed Traditional/Nontraditional",
                                                3:"Associate's Colleges: High Transfer-High Nontraditional",
                                                4:"Associate's Colleges: Mixed Transfer/Career & Technical-High Traditional",
                                                5:"Associate's Colleges: Mixed Transfer/Career & Technical-Mixed Traditional/Nontraditional",
                                                6:"Associate's Colleges: Mixed Transfer/Career & Technical-High Nontraditional",
                                                7:"Associate's Colleges: High Career & Technical-High Traditional",
                                                8:"Associate's Colleges: High Career & Technical-Mixed Traditional/Nontraditional",
                                                9:"Associate's Colleges: High Career & Technical-High Nontraditional",
                                                10:"Special Focus Two-Year: Health Professions",
                                                11:"Special Focus Two-Year: Technical Professions",
                                                12:"Special Focus Two-Year: Arts & Design",
                                                13:"Special Focus Two-Year: Other Fields",
                                                14:"Baccalaureate/Associate's Colleges: Associate's Dominant",
                                                15:"Doctoral Universities: Very High Research Activity",
                                                16:"Doctoral Universities: High Research Activity",
                                                17:"Doctoral/Professional Universities",
                                                18:"Master's Colleges & Universities: Larger Programs",
                                                19:"Master's Colleges & Universities: Medium Programs",
                                                20:"Master's Colleges & Universities: Small Programs",
                                                21:"Baccalaureate Colleges: Arts & Sciences Focus",
                                                22:"Baccalaureate Colleges: Diverse Fields",
                                                23:"Baccalaureate/Associate's Colleges: Mixed Baccalaureate/Associate's",
                                                24:"Special Focus Four-Year: Faith-Related Institutions",
                                                25:"Special Focus Four-Year: Medical Schools & Centers",
                                                26:"Special Focus Four-Year: Other Health Professions Schools",
                                                27:"Special Focus Four-Year: Research Institution",
                                                28:"Special Focus Four-Year: Engineering and Other Technology-Related Schools",
                                                29:"Special Focus Four-Year: Business & Management Schools",
                                                30:"Special Focus Four-Year: Arts, Music & Design Schools",
                                                31:"Special Focus Four-Year: Law Schools",
                                                32:"Special Focus Four-Year: Other Special Focus Institutions",
                                                33:"Tribal Colleges and Universities"})

carnegie_df['hbcu'] = carnegie_df.hbcu.replace({1:'HBCU', 2:'not HBCU'})
carnegie_df['tribal'] = carnegie_df.tribal.replace({1:'tribal', 2:'not tribal'})
carnegie_df['hsi'] = carnegie_df.hsi.replace({0:'not HSI', 1:'HSI'})
carnegie_df['womens'] = carnegie_df.womens.replace({0:'not womens', 1:'Womens'})

hbcu = carnegie_df[carnegie_df.hbcu == 'HBCU']
hsi = carnegie_df[carnegie_df.hsi == 'HSI']
womens = carnegie_df[carnegie_df.womens == 'Womens']
tribal = carnegie_df[carnegie_df.tribal == 'tribal']

writer = pd.ExcelWriter('../../handcoding/carnegie_flags.xlsx')

hbcu.to_excel(writer,index=False, sheet_name='hbcu')
tribal.to_excel(writer,index=False, sheet_name='tribal',)
hsi.to_excel(writer,index=False, sheet_name='hsi')
womens.to_excel(writer,index=False, sheet_name='womens')

writer.save()
