==> code/1_inst_cleaning/1_institutional_impact.ipynb <==
%config Completer.use_jedi = False

import pandas as pd
import numpy as np


address = pd.read_csv('/data/datasets/WOS/US/Adresses_race.txt', delimiter='\t')

address

institutional_type = pd.read_csv('../../data/institutions_us.csv')
institutions_impact = pd.read_excel('/data/datasets/WOS/US/institutions_impact.xlsx', names = ['Institution','country', 'npapers','avg_citations'])

institutions_impact['impact'] = institutions_impact.npapers * institutions_impact.avg_citations

institutions_impact_us = institutions_impact[institutions_impact.country == 'United States of America']

df_institutions = address.merge(institutions_impact_us, how='left', on='Institution').merge(institutional_type, how='left', on='Institution')

df_institutions.to_pickle('/data/datasets/WOS/US/Adresses_impact.p')
