==> code/1_inst_cleaning/2_create_handcoding_file.ipynb <==
%config Completer.use_jedi = False

import pandas as pd
import pickle
from tqdm.notebook import tqdm
import numpy as np
import re
import string

institutional_type = pd.read_csv('../../data/institutions_us.csv')
institutions_impact = pd.read_excel('/data/datasets/WOS/US/institutions_impact.xlsx', names = ['Institution','country', 'npapers','avg_citations'])

address = pd.read_pickle('/data/datasets/WOS/US/Adresses_impact.p')

grid_df = pd.read_csv('../../data/grid-2021-09-16/grid.csv')
grid_type = pd.read_csv('../../data/grid-2021-09-16/full_tables/types.csv')

## updated to new 2021 classification
# carnegie_df = pd.read_csv('../../data/carnegie/Data_1-7-2022---920.csv')
# carnegie_labels = pd.read_csv('../../data/carnegie/ValueLabels_1-7-2022---920.csv')
# carnegie_df = carnegie_df[['UnitID','Institution Name']]
carnegie_df = pd.read_excel('../../data/carnegie/CCIHE2021-PublicData.xlsx',sheet_name='Data')
carnegie_df = carnegie_df[['unitid','name']]
carnegie_df.columns = ['carnegie_id','carnegie_name']

carnegie_df['carnegie_id'] = carnegie_df.carnegie_id.astype("category")

institution_count = address.Institution.value_counts().reset_index()
institution_count.columns=['name','n']

institution_count['name'] = institution_count.name.str.upper()

institution_count = institution_count.merge(institutional_type, left_on='name', right_on='Institution', how='left')

institution_count.drop(columns='Institution',inplace=True)

institution_count.sector.value_counts()

## Automatic cleaning

def rename_to_greater(institution_count,to_omit, threshold):
    
    ic = institution_count.copy()
    names = to_omit

    ic['new_name'] = ic.name

    for index, row in tqdm(ic[ic.n>threshold].iterrows()): # I only clean names of institutions with n>100, 5K institutions
        if row.new_name in names:
            continue
        else:
            name = row['new_name']
            sector = row['sector']
            ic_names = ic.new_name.values
            matched_names = ic_names[[bool(re.search(pattern=f'(^|\s|-){name}($|\s|-)', string=n)) for n in ic_names]]
            ic.loc[ic.name.isin(matched_names),'new_name'] = name
            ic.loc[ic.name.isin(matched_names),'sector'] = sector
            names += [name]

    return ic

states = ['ALASKA', 'ALABAMA', 'ARKANSAS', 'AMERICAN SAMOA', 'ARIZONA', 'CALIFORNIA', 'COLORADO', 'CONNECTICUT', 'COLUMBIA', 'DELAWARE', 'FLORIDA', 'GEORGIA', 'GUAM', 'HAWAII', 'IOWA', 'IDAHO', 'ILLINOIS', 'INDIANA', 'KANSAS', 'KENTUCKY', 'LOUISIANA', 'MASSACHUSETTS', 'MARYLAND', 'MAINE', 'MICHIGAN', 'MINNESOTA', 'MISSOURI', 'MISSISSIPPI', 'MONTANA', 'NORTH CAROLINA', 'NORTH DAKOTA', 'NEBRASKA', 'NEW HAMPSHIRE', 'NEW JERSEY', 'NEW MEXICO', 'NEVADA', 'NEW YORK', 'OHIO', 'OKLAHOMA', 'OREGON', 'PENNSYLVANIA','PENN', 'PUERTO RICO', 'RHODE ISLAND', 'SOUTH CAROLINA', 'SOUTH DAKOTA', 'TENNESSEE', 'TEXAS', 'UTAH', 'VIRGINIA', 'VIRGIN ISLANDS', 'VERMONT', 'WASHINGTON', 'WISCONSIN', 'WEST VIRGINIA', 'WYOMING']

state_unis = ['UNIV-' + state for state in states] + [state + '-UNIV' for state in states]

to_omit = ['LLC','MED-CTR','RES-INST', 'CHILDRENS-HOSP','USA','SCH-MED','VET-ADM','UNIV-CALIF','UNC','UNIV-HOSP'] + state_unis

ic = rename_to_greater(institution_count, to_omit,1000)

ic.new_name.value_counts()[:40]

new_ic= ic.groupby('new_name').n.sum().reset_index().sort_values('n',ascending=False)

print(f"we retrieve {round(new_ic[new_ic.n>=500].n.sum()/new_ic.n.sum()*100,2)}% of the database,")
print(f"represented by {len(new_ic[new_ic.n>=500])} institutions,")
print(f"which are the {round(len(new_ic[new_ic.n>=500])/len(new_ic)*100,3)}% of the institutions.")



large_institutions = new_ic[new_ic.n>=500]

large_institutions_all = ic[ic.new_name.isin(large_institutions.new_name)]

ic.groupby('sector').n.sum()/ic.groupby('sector').n.sum().sum()

institutions_to_clean = large_institutions_all[large_institutions_all.sector.isin(['university','hospital'])]

institutions_to_clean.sort_values(['new_name','n'],ascending=[True,False])

sorter = institutions_to_clean.new_name.unique()

institutions_to_clean.loc[:,'new_name'] = institutions_to_clean.new_name.astype("category")

institutions_to_clean.loc[:,'new_name'] = institutions_to_clean.new_name.cat.set_categories(sorter)


institutions_to_clean = institutions_to_clean.sort_values(['new_name','n'],ascending=[True,False])

### Partial match with GRID

grid_us = grid_df[grid_df.Country == 'United States']

grid_us = grid_us.merge(grid_type,left_on='ID', right_on='grid_id')

def standardize_name(name):
    name = name.translate(str.maketrans('','',string.punctuation)) # remove punctuation
    name = re.sub('the|of','',name,flags=re.IGNORECASE)
    name = re.sub(' +', ' ', name) #remove multiple white space
    name = name.upper()
    name = name.replace(' ','-')
    name = name.replace('UNIVERSITY','UNIV')
    name = name.replace('INSTITUTE','INST')
    name = name.replace('DIVISION','DIV')
    name = name.replace('CLINIQUE','CLI')
    name = name.replace('MEDICAL','MED')    
    name = name.replace('CENTER','CTR')  
    name = name.replace('RESEARCH','RES')  
    name = name.replace('RESEARCH','RES')  
    name = name.replace('COLLEGE','COLL')  
    name = name.replace('ACADEMY','ACAD')      
    name = name.replace('CALIFORNIA','CALIF')      
    name = name.replace('PENNSYLVANIA','PENN')
    return name

text = 'University of California, Los Angeles'
standardize_name(text)

grid_us.loc[:,'standard_name'] = grid_us.Name.apply(standardize_name)

grid_us_merge = grid_us[['grid_id','Name','standard_name']]
grid_us_merge.columns = ['grid_id','grid_name','grid_standard_name']

def double_merge_grid(institutions_to_clean, grid_us_merge):

    partial_match = institutions_to_clean.merge(grid_us_merge, left_on='new_name', right_on='grid_standard_name', how='left')

    partial_match.drop(columns='grid_standard_name',inplace=True)

    not_matched = partial_match[partial_match.grid_id.isnull()]

    not_matched.drop(columns=['grid_id', 'grid_name'],inplace=True)

    second_match = not_matched.merge(grid_us_merge, left_on='name', right_on='grid_standard_name', how='inner')
    second_match.drop(columns='grid_standard_name',inplace=True)

    partial_match.loc[partial_match.name.isin(second_match.name),['grid_id','grid_name']] = second_match[['grid_id','grid_name']].values

    return partial_match

partial_match = double_merge_grid(institutions_to_clean, grid_us_merge)

### Partial match with Carnegie

carnegie_df.loc[:,'carnegie_standard_name'] = carnegie_df.carnegie_name.apply(standardize_name)

def double_merge_carnegie(partial_match, carnegie_df):

    partial_match = partial_match.merge(carnegie_df, left_on='new_name', right_on='carnegie_standard_name', how='left')

    partial_match.drop(columns='carnegie_standard_name',inplace=True)

    not_matched = partial_match[partial_match.carnegie_id.isnull()]

    not_matched.drop(columns=['carnegie_id', 'carnegie_name'],inplace=True)

    second_match = not_matched.merge(carnegie_df, left_on='name', right_on='carnegie_standard_name', how='inner')
    second_match.drop(columns='carnegie_standard_name',inplace=True)

    partial_match.loc[partial_match.name.isin(second_match.name),['carnegie_id', 'carnegie_name']] = second_match[['carnegie_id', 'carnegie_name']].values

    return partial_match

matched_institutions = double_merge_carnegie(partial_match, carnegie_df)

matched_institutions

matched_institutions.to_excel('../../handcoding/institutions.xlsx')
