==> code/1_inst_cleaning/4_institutions_impact.ipynb <==
## Legacy code

new goals:

1. Valuative classification: proportion of papers in top 1%/5% most cited)
2. Perceptual classification: US news report. 

%config Completer.use_jedi = False

import pandas as pd
import pickle
from tqdm.notebook import tqdm
import numpy as np
from pandas._libs.lib import is_integer


from libs.LastNamesInference import LastNamesInference

def split_by_gender(df_race):
    df_race = df_race.copy()
    df_race['gender'] = df_race.gender.str.upper()

    boolean_mask_M = df_race.gender == 'M'
    boolean_mask_F = df_race.gender == 'F'

    race_dist_M = df_race.loc[boolean_mask_M,['white','hispanic','black','asian']]
    race_dist_F = df_race.loc[boolean_mask_F,['white','hispanic','black','asian']]

    race_dist_M.columns = race_dist_M.columns + '_M'
    race_dist_F.columns = race_dist_F.columns + '_F'
    race_dist_MF = pd.concat([race_dist_M,race_dist_F]).fillna(0).sort_index()
    df_race = df_race.join(race_dist_MF,how='inner' )
    return df_race


def infer_race(us_papers, df=None):
    papers = us_papers.drop_duplicates('cluster_ID')
    lni = LastNamesInference(names = papers.nom)
    if df is None:
        df = us_papers.copy()
    tqdm.pandas(desc="inferring race from lastnames")
    lastname_race_dist = df.progress_apply(lambda x: lni.get_name_dist(lastname=x.nom), axis=1)
    df[lni.prob_order] = pd.DataFrame(lastname_race_dist.to_list(), index=df.index)
    df = split_by_gender(df)
    return df

def weighted_qcut(values, weights, q, **kwargs):
    'Return weighted quantile cuts from a given series, values.'
    if is_integer(q):
        quantiles = np.linspace(0, 1, q + 1)
    else:
        quantiles = q
    order = weights.iloc[values.argsort()].cumsum()
    bins = pd.cut(order / order.iloc[-1], quantiles, **kwargs)
    return bins.sort_index()

def weighted_qcut_labels(df, var):
    lims = df.groupby('{}_Q'.format(var)).agg({var:[lambda x: round(min(x),ndigits=2),lambda x: round(max(x),ndigits=2)]}).droplevel(axis=1,level=1)
    lims.columns = ['min','max']
    lims['label'] = lims.apply(lambda x: '({}, {})'.format(x['min'], x['max']),axis=1)
    lims = lims.drop(columns=['min','max'])
    qlabels = df['{}_Q'.format(var)].map(lims['label'])
    return qlabels

# us_papers = pd.read_csv('/data/datasets//WOS/US/US_papers.txt')
# us_papers.to_pickle('/data/datasets//WOS/US/US_papers.p')
us_papers = pd.read_pickle('/data/datasets//WOS/US/US_papers.p')

us_papers_citations = pd.read_csv('/data/datasets/WOS/US/us_papers_citations.csv', delimiter='\t')

us_papers_citations = us_papers_citations.rename(columns={'ID_Art':'id_art'})

institutions_impact = pd.read_excel('../../data/impact.xlsx')
#institutions_impact = pd.read_excel('/data/datasets/WOS/US/institutions_impact.xlsx', names = ['institution','country', 'npapers','avg_citations'])
#address = pd.read_csv('/data/datasets/WOS/US/Adresses_race.txt', delimiter='\t')

institutions_impact = institutions_impact[pd.to_numeric(institutions_impact['carnegie_id'], errors='coerce').notnull()]

institutions_impact['carnegie_id'] = institutions_impact.carnegie_id.astype(str)

handcoding_df = pd.read_excel('../../data/institutions_20220329.xlsx',sheet_name='Sheet1',converters={'index':int,'carnegie_id':int})
handcoding_df['carnegie_id'] = handcoding_df.carnegie_id.astype('str')
handcoding_flags = pd.read_excel('../../handcoding/carnegie_flags.xlsx',sheet_name='to_clean',converters={'real_unitid':str})

crosswalk_wos_carnegie = handcoding_df[['name','carnegie_id']]

_ = handcoding_flags[['name_wos','real_unitid']].dropna().drop_duplicates().\
rename(columns={'name_wos':'name','real_unitid':'carnegie_id'})

crosswalk_wos_carnegie = crosswalk_wos_carnegie.append(_)

address_clean= pd.read_pickle('/data/datasets/WOS/US/address_clean_carnegie.p')

### proprotion of articles in the top x% most cited

df_topcr = us_papers_citations.merge(address_clean[['id_art','carnegie_id']].drop_duplicates(),on='id_art')

agg_topcr = df_topcr.groupby('carnegie_id').agg({'id_art':len, 'Top10CR':sum,'Top5CR':sum,'Top1CR':sum})

agg_topcr['Top10CR'] = agg_topcr.Top10CR/agg_topcr.id_art
agg_topcr['Top5CR'] = agg_topcr.Top5CR/agg_topcr.id_art
agg_topcr['Top1CR'] = agg_topcr.Top1CR/agg_topcr.id_art

agg_topcr = agg_topcr.rename(columns={'id_art':'ncitable_docs'})

# I use weighted quartiles, to consider the size of institutions

agg_topcr['Top10CR_Q'] = weighted_qcut(agg_topcr['Top10CR'], agg_topcr['ncitable_docs'], 3, labels=False)
agg_topcr['Top5CR_Q'] = weighted_qcut(agg_topcr['Top5CR'], agg_topcr['ncitable_docs'], 3, labels=False)
agg_topcr['Top1CR_Q'] = weighted_qcut(agg_topcr['Top1CR'], agg_topcr['ncitable_docs'], 3, labels=False)

# agg_topcr['Top10CR_Q'] = pd.qcut(agg_topcr.Top10CR, 4)
# agg_topcr['Top5CR_Q'] = pd.qcut(agg_topcr.Top5CR, 4)
# agg_topcr['Top1CR_Q'] = pd.qcut(agg_topcr.Top1CR, 4)

agg_topcr['Top10CR_Q'] = weighted_qcut_labels(agg_topcr, 'Top10CR')
agg_topcr['Top5CR_Q'] = weighted_qcut_labels(agg_topcr, 'Top5CR')
agg_topcr['Top1CR_Q'] = weighted_qcut_labels(agg_topcr, 'Top1CR')

agg_topcr.Top5CR_Q.value_counts()

address_clean = address_clean.merge(agg_topcr, on = 'carnegie_id')

## impact

# impact_df = crosswalk_wos_carnegie.merge(institutions_impact, left_on='name',right_on='institution')\
# [['carnegie_id','npapers','avg_citations']].copy()

# impact_df = impact_df[impact_df.carnegie_id!='nan']

# impact_df['impact'] = impact_df.npapers*impact_df.avg_citations

# impact_df = impact_df.groupby('carnegie_id').agg({'npapers':sum, 'impact':sum})

# impact_df['avg_citations'] = impact_df.impact/impact_df.npapers # I reconstruct the average citations as the weighted average

impact_df = institutions_impact.copy()

impact_df = impact_df[impact_df.carnegie_id.isin(address_clean.carnegie_id.unique())]

# I use weighted quartiles, to consider the size of institutions

impact_df['npapers_Q'] = weighted_qcut(impact_df['npapers'], impact_df['npapers'], 3, labels=False)
impact_df['impact_Q'] = weighted_qcut(impact_df['impact'], impact_df['npapers'], 3, labels=False)
impact_df['avg_citations_Q'] = weighted_qcut(impact_df['avg_citations'], impact_df['npapers'], 3, labels=False)

impact_df['npapers_Q'] = weighted_qcut_labels(impact_df, 'npapers')
impact_df['impact_Q'] = weighted_qcut_labels(impact_df, 'impact')
impact_df['avg_citations_Q'] = weighted_qcut_labels(impact_df, 'avg_citations')

impact_df.npapers_Q.value_counts()

# impact_df['npapers_Q'] = pd.qcut(impact_df.npapers, 4)
# impact_df['impact_Q'] = pd.qcut(impact_df.impact, 4)
# impact_df['avg_citations_Q'] = pd.qcut(impact_df.avg_citations, 4)

address_clean = address_clean.merge(impact_df, on='carnegie_id', how='left')

address_clean.npapers_Q.value_counts()

address_clean.avg_citations_Q.value_counts()

address_clean.avg_citations_Q.value_counts()

address_clean.loc[address_clean.avg_citations_Q=='(1.48, 1.74)','avg_citations'].max()

address_clean.loc[address_clean.avg_citations_Q=='(1.77, 4.07)','avg_citations'].min()

address_clean.to_pickle('/data/datasets/WOS/US/address_clean.p')
