==> code/1_inst_cleaning/5_NewsReport.ipynb <==
%config Completer.use_jedi = False

import pandas as pd
import pickle
from tqdm.notebook import tqdm
import numpy as np
import re
import string

Save dataset for handcoding

address_clean = pd.read_pickle('/data/datasets/WOS/US/address_clean.p')

#for repeated runs
#address_clean = address_clean.drop(columns=['usnr_rank','usnr_rank_cat'])

carnegie_universities = address_clean[['grid_id','carnegie_id','carnegie_name']].drop_duplicates('carnegie_id')

carnegie_universities = carnegie_universities.merge(address_clean.carnegie_id.value_counts().to_frame(name='n'), how='left',left_on='carnegie_id',right_index=True)

carnegie_universities = carnegie_universities.sort_values('n',ascending=False)

#carnegie_universities.to_excel('../../handcoding/universities_newsreport_for_handcoding.xlsx', index=False)

Load top 100 universities

us_newsreport = pd.read_excel('../../handcoding/USNEWS Ranking + Carnegie .xlsx', names=['usnr_rank','institution_name','carnegie_id'])

#us_newsreport['usnr_rank'] = us_newsreport.usnr_rank.astype(int)
us_newsreport['carnegie_id'] = us_newsreport.carnegie_id.astype(int).astype('str')

address_clean = address_clean.merge(us_newsreport[['usnr_rank','carnegie_id']], on = 'carnegie_id', how='left')

# I assign the ones that are not in the top 100 to 999
address_clean.loc[address_clean.usnr_rank.isnull(),'usnr_rank'] = 999

test =pd.cut(address_clean.usnr_rank,bins=[0,10,100,999], labels=['top_10','top_100','not_top'])

address_clean['usnr_rank_cat'] = pd.cut(address_clean.usnr_rank,bins=[0,10,100,999], labels=['top_10','top_100','not_top'])

address_clean.usnr_rank_cat.value_counts()

address_clean.to_pickle('/data/datasets/WOS/US/address_clean.p')

#address_clean = pd.read_pickle('/data/datasets/WOS/US/address_clean.p')

institutions = address_clean[['carnegie_id','carnegie_name','avg_citations','avg_citations_Q','usnr_rank','usnr_rank_cat','selindex']].drop_duplicates()

institutions.to_csv('../../data/institutions_prestige_metrics.csv', index=False)

address_clean[address_clean.usnr_rank_cat.isnull()]

institutions[institutions.usnr_rank_cat.isnull()]
